Trying to make BoVeX run from arbitrary directories. Renamed some ancient Util functions, which meant fixing a bunch of old programs
git-svn-id: file:///home/runner/work/tom7misc-sync-action/tom7misc-sync-action/upstream/trunk@5882 c7e11af0-4c7a-11f1-9edd-117cb6f09d55

+++ cc-lib/do-not-optimize_test.cc
@@ -14,7 +14,7 @@ inline const char *Function(const char *str) {
 
 static std::string Reverse(std::string s) {
   char *buf = (char*)malloc(s.size());
-  for (int i = 0; i < s.size(); i++)
+  for (int i = 0; i < (int)s.size(); i++)
     buf[i] = s[s.size() - 1 - i];
   return std::string(buf, s.size());
 }
